Confidence interval of the total under a Simple Random Sampling framework

Confidence interval of the total under a Simple Random Sampling framework

ct_srs(C, E, t_est, sd_exp, n_real, sd_est, parameter = FALSE, N)

Arguments

C

Level of confidence. (0 <= C <= 1)

E

Sampling error. (E > 0)

t_est

Total estimated from the sample

sd_exp

Expected standard deviation (sd_exp > 0)

n_real

Real sample size (n_real > 0)

sd_est

Sample standar deviation (sd_est > 0)

parameter

Type TRUE if you do know the populations sd, type FALSE (default) if it is an estimate.

N

A positive integer indicating the number of elements in the population.

Value

The function returns the interval of confidence of the population total, and information regarding the sufficiency of the sample size.

Examples

ct_srs(C = 0.95, E = 1000, t_est = 3500, sd_exp = 4.1, n_real = 87, sd_est = 4.1, N = 1200)
#> With 95% confidence, the population total is between 2502.92 and 4497.08 
#> Your sample size: 87 
#> Needed sample size: 87 
#> Your sample size seems to be sufficient.
ct_srs(C = 0.95, E = 1000, t_est = 3500, sd_exp = 4.1,
       n_real = 87, sd_est = 6, parameter = TRUE, N = 1200)
#> With 95% confidence, the population total is between 2042.331 and 4957.669 
#> Your sample size: 87 
#> Needed sample size: 171 
#> Consider increasing the sample size if you want to reduce the width of the interval