Confidence interval of the mean under a Simple Random Sampling framework

Confidence interval of the mean under a Simple Random Sampling framework

cx_srs(C, E, x_est, sd_exp, n_real, sd_est, parameter = FALSE, N = Inf)

Arguments

C

Level of confidence. (0 <= C <= 1)

E

Sampling error. (E > 0)

x_est

Sample mean

sd_exp

Expected standard deviation (s_exp > 0)

n_real

Real sample size (n_real > 0)

sd_est

Sample standar deviation (s_est > 0)

parameter

Type TRUE if you do know the populations sd, type FALSE (default) if it is an estimate.

N

A positive integer indicating the number of elements in the population. By default, infinite.

Value

The function returns the interval of confidence of the population mean, and information regarding the sufficiency of the sample size.

Examples

cx_srs(C = 0.95, E = 50, x_est = 3500, sd_est = 200, n_real = 250, sd_exp = 200)
#> With 95% confidence, the population mean is between 3475.208 and 3524.792 
#> Your sample size: 250 
#> Needed sample size: 62 
#> Your sample size seems to be sufficient.
cx_srs(C = 0.95, E = 50, x_est = 3500, sd_est = 400, n_real = 400, sd_exp = 200)
#> With 95% confidence, the population mean is between 3460.801 and 3539.199 
#> Your sample size: 400 
#> Needed sample size: 62 
#> Your sample size seems to be sufficient.
cx_srs(C = 0.95, E = 50, x_est = 3500, sd_est = 400,
        n_real = 250, sd_exp = 200, parameter = TRUE, N = 5000)
#> With 95% confidence, the population mean is between 3451.667 and 3548.333 
#> Your sample size: 250 
#> Needed sample size: 61 
#> Your sample size seems to be sufficient.